changed error dialogs

+++ restore.sh
@@ -2233,12 +2233,11 @@ elif [ "$BRinterface" = "dialog" ]; then
       ( $BRarchiver tf /mnt/target/fullbackup 2>&1 || touch /tmp/tar_error ) |
       tee /tmp/filelist | while read ln; do a=$(( a + 1 )) && echo -en "\rReading archive: $a Files "; done | dialog --progressbox 3 40
       sleep 1
-      ( if [ -f /tmp/tar_error ]; then
+      if [ -f /tmp/tar_error ]; then
         rm /tmp/tar_error
-        echo -e "Error reading archive"
         rm /mnt/target/fullbackup
-        sleep 2
-      fi ) | dialog --progressbox 3 30
+        dialog --title "Error" --msgbox "Error reading archive." 5 26
+      fi
     fi
 
     while [ ! -f /mnt/target/fullbackup ]; do
@@ -2324,12 +2323,11 @@ elif [ "$BRinterface" = "dialog" ]; then
         ( $BRarchiver tf /mnt/target/fullbackup 2>&1 || touch /tmp/tar_error ) |
         tee /tmp/filelist | while read ln; do a=$(( a + 1 )) && echo -en "\rReading archive: $a Files "; done | dialog --progressbox 3 40
         sleep 1
-        ( if [ -f /tmp/tar_error ]; then
+        if [ -f /tmp/tar_error ]; then
           rm /tmp/tar_error
-          echo -e "Error reading archive"
           rm /mnt/target/fullbackup
-          sleep 2
-        fi ) | dialog --progressbox 3 30
+          dialog --title "Error" --msgbox "Error reading archive." 5 26
+        fi
       fi
     done
   fi
